Refactor Generator component with SpringBootGenerator
