+++ src/main/webapp/app/springboot/primary/spring-boot-generator/index.ts
@@ -1,0 +1,4 @@
+import SpringBootGeneratorComponent from './SpringBootGenerator.component';
+import SpringBootGeneratorVue from './SpringbootGenerator.vue';
+
+export { SpringBootGeneratorComponent, SpringBootGeneratorVue };
